Adding the ability to control looping
I like the simplicity of this library. I do though think it could use the ability for a developer to set if it never loops, always loops or loops n times. So I added some code that if you add the attribute as such:

It will loop for ever (it's also the default)
```
data-loop="true"
```

It will not loop at all, it will just type out once.
```
data-loop="false"
```

It will loop 4 times
```
data-loop="4"
```

This unfortunately only works with non-delimited words, mainly because multiple delimitations was just not a use case for me. I will submit another change once I get to that.